Add error reporting to Antlr parsers
If you look at antlrlexerwrapper.rs, you'll see we basically needed to
copy entirely the error recovery class. This is due to it assuming that
you can build tokens simply from a string, and also assuming that the
errors would come from a single location. This of course is not the case
for us, so we need to do a bit more work. I also had to completely
circumbent the error listener machinery, as it is not capable of
reporting from multiple files.

A bit of a hack, but at least it's a self contained hack.