Pass anotated dependency nodes.
We'll check that what we truly parse is what we naively parsed. We don't
run the preprocessor while making the dependency tree in order to avoid
the chicken and egg problem of the importable headers. During the true
parsing, we can now check that the dependencies match correctly, and
that the user truly didn't use macros inside the module directives.