Implement some tests for unqualified name resolution
Also, found a bug caused by the use of unsafe. I was using referenes to
elements in a vector, that I had marked static, and appending stuff to
said vector. It is easy to see where the mistake was.

I've removed almost all the unsafe code, except for the unsafe
allocator. Its use is reduced and thought out, so I want to believe we
won't be bitten by it.